Xappy-Oil/backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# AI/LLM
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
tiktoken>=0.6.0

# Environment
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# File handling
Pillow>=10.2.0

# Audio processing (requires ffmpeg system dependency)
pydub>=0.25.0

# PDF/Excel export
reportlab>=4.0.0
openpyxl>=3.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
tenacity>=8.2.0
python-dateutil>=2.8.0
pytz>=2024.1
orjson>=3.9.0
